init table utils, init endless parsedata loop fix
